Migrate ViewShape/Style/Data and View to dc_bundle
Removes the now-empty toolkit_schema

Fixes Migrate ViewStyle/View to dc_bundle #1456
Fixes Migrate ViewShape/Data to dc_bundle #1455
