+++ crates/dc_bundle/src/legacy_definition/view/view.rs
@@ -14,7 +14,16 @@
  * limitations under the License.
  */
 
+use crate::legacy_definition::element::geometry::Rectangle;
+use crate::legacy_definition::element::reactions::{FrameExtras, Reaction};
+use crate::legacy_definition::element::view_shape::ViewShape;
+use crate::legacy_definition::layout::positioning::OverflowDirection;
+use crate::legacy_definition::view::component::ComponentInfo;
+use crate::legacy_definition::view::text_style::StyledTextRun;
+use crate::legacy_definition::view::view_style::ViewStyle;
 use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
+use std::sync::atomic::AtomicU16;
 
 /// This enum may be used as a hint by the DesignCompose renderer implementation
 /// to determine if it is important for the content to be rendered identically on different platforms.
@@ -22,4 +31,147 @@ use serde::{Deserialize, Serialize};
 pub enum RenderMethod {
     None,
     PixelPerfect,
+}
+
+/// Details that are unique to each view type.
+#[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
+pub enum ViewData {
+    Container { shape: ViewShape, children: Vec<View> },
+    Text { content: String, res_name: Option<String> },
+    StyledText { content: Vec<StyledTextRun>, res_name: Option<String> },
+}
+
+/// This struct contains information for scrolling on a frame. It combines the
+/// scroll overflow direction, which comes from a frame, and the bool
+/// paged_scrolling, which comes from the vsw-extended-layout plugin.
+#[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
+pub struct ScrollInfo {
+    pub overflow: OverflowDirection,
+    pub paged_scrolling: bool,
+}
+
+impl Default for ScrollInfo {
+    fn default() -> Self {
+        ScrollInfo { overflow: OverflowDirection::None, paged_scrolling: false }
+    }
+}
+
+/// Represents a toolkit View (like a Composable).
+#[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
+pub struct View {
+    // unique numeric id (not replicated within one DesignCompose file), which is used by the
+    // renderer when building a layout tree, or when mapping from an integer to a View.
+    pub unique_id: u16,
+    // id doesn't exist in vsw toolkit, but is useful for tracing from Figma node to output
+    pub id: String,
+    // name doesn't exist in vsw toolkit, but is also useful for tracing.
+    pub name: String,
+    // information on the component that this view is an instance of (if any).
+    pub component_info: Option<ComponentInfo>,
+    // interactivity
+    pub reactions: Option<Vec<Reaction>>,
+    // interactivity related frame properties
+    pub frame_extras: Option<FrameExtras>,
+    // overflow (scrolling) data
+    pub scroll_info: ScrollInfo,
+    // common style
+    pub style: ViewStyle,
+    pub data: ViewData,
+    pub design_absolute_bounding_box: Option<Rectangle>,
+    pub render_method: RenderMethod,
+    // a hash of variable collection id -> mode id if variable modes are set on this view
+    pub explicit_variable_modes: Option<HashMap<String, String>>,
+}
+
+impl View {
+    fn next_unique_id() -> u16 {
+        static COUNTER: AtomicU16 = AtomicU16::new(0);
+        COUNTER.fetch_add(1, std::sync::atomic::Ordering::Relaxed)
+    }
+    pub fn new_rect(
+        id: &String,
+        name: &String,
+        shape: ViewShape,
+        style: ViewStyle,
+        component_info: Option<ComponentInfo>,
+        reactions: Option<Vec<Reaction>>,
+        scroll_info: ScrollInfo,
+        frame_extras: Option<FrameExtras>,
+        design_absolute_bounding_box: Option<Rectangle>,
+        render_method: RenderMethod,
+        explicit_variable_modes: Option<HashMap<String, String>>,
+    ) -> View {
+        View {
+            unique_id: View::next_unique_id(),
+            id: id.clone(),
+            name: name.clone(),
+            component_info,
+            reactions,
+            style,
+            frame_extras,
+            scroll_info,
+            data: ViewData::Container { shape, children: vec![] },
+            design_absolute_bounding_box,
+            render_method,
+            explicit_variable_modes,
+        }
+    }
+    pub fn new_text(
+        id: &String,
+        name: &String,
+        style: ViewStyle,
+        component_info: Option<ComponentInfo>,
+        reactions: Option<Vec<Reaction>>,
+        text: &str,
+        text_res_name: Option<String>,
+        design_absolute_bounding_box: Option<Rectangle>,
+        render_method: RenderMethod,
+        explicit_variable_modes: Option<HashMap<String, String>>,
+    ) -> View {
+        View {
+            unique_id: View::next_unique_id(),
+            id: id.clone(),
+            name: name.clone(),
+            component_info,
+            reactions,
+            style,
+            frame_extras: None,
+            scroll_info: ScrollInfo::default(),
+            data: ViewData::Text { content: text.into(), res_name: text_res_name },
+            design_absolute_bounding_box,
+            render_method,
+            explicit_variable_modes,
+        }
+    }
+    pub fn new_styled_text(
+        id: &String,
+        name: &String,
+        style: ViewStyle,
+        component_info: Option<ComponentInfo>,
+        reactions: Option<Vec<Reaction>>,
+        text: Vec<StyledTextRun>,
+        text_res_name: Option<String>,
+        design_absolute_bounding_box: Option<Rectangle>,
+        render_method: RenderMethod,
+    ) -> View {
+        View {
+            unique_id: View::next_unique_id(),
+            id: id.clone(),
+            name: name.clone(),
+            style,
+            component_info,
+            reactions,
+            frame_extras: None,
+            scroll_info: ScrollInfo::default(),
+            data: ViewData::StyledText { content: text, res_name: text_res_name },
+            design_absolute_bounding_box,
+            render_method,
+            explicit_variable_modes: None,
+        }
+    }
+    pub fn add_child(&mut self, child: View) {
+        if let ViewData::Container { children, .. } = &mut self.data {
+            children.push(child);
+        }
+    }
 }
